+++ CHANGELOG.md
@@ -2,14 +2,24 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Semantic Versioning](https://semver.org/).
 
+## version: 1.0.2   (11/12/2024)
+### Added
+
+- Updated Mysql Configurations.
+- Updated Redis Configurations.
+- Added Automation for server setup.
+- Added Inku, your very own assistant in Incognito-Vault. 
+- Updated Project Directory.
+- Added Folder `Database` to the directory structure.
+
 ## version: 1.0.1  
 ### Added
 
 - Added API Endpoint.
 - Updated the Base Html File.
-- Enchanced the Project updating script.
-- Updated Readme.
- File
+- Enhanced the Project updating script.
+- Updated Readme File.
+
 ## version: 1.0.0  03-12-2024
 ### Added
 - Added Versioning support.